% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinep.r
\name{combinep_three}
\alias{combinep_three}
\title{combinep_three}
\usage{
combinep_three(p1, p2, p3)
}
\arguments{
\item{p1}{the p-values or corrected p-values}

\item{p2}{the p-values or corrected p-values}

\item{p3}{the p-values or corrected p-values}
}
\value{
A numeric vector of z_scores
}
\description{
`combinep_three` combine three kinds of p-values,then,calculate z-score for them.
}
\examples{
exp.p<-GetExampleData("exp.p")
meth.p<-GetExampleData("meth.p")
cnv.p<-GetExampleData("cnv.p")
\donttest{combinep_three(exp.p,meth.p,cnv.p)}
}
