% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinep.r
\name{combinep_two}
\alias{combinep_two}
\title{combinep_two}
\usage{
combinep_two(p1, p2)
}
\arguments{
\item{p1}{A numeric vector of p-values or corrected p-values}

\item{p2}{A numeric vector of p-values or corrected p-values}
}
\value{
A numeric vector of z_scores
}
\description{
`combinep_two` combine two kinds of p-values,then,calculate z-score for them.
}
\examples{
exp.p<-GetExampleData("exp.p")
meth.p<-GetExampleData("meth.p")
combinep_two(exp.p,meth.p)
}
