% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetExampleData.R
\name{GetExampleData}
\alias{GetExampleData}
\title{Get the example data}
\usage{
GetExampleData(exampleData)
}
\arguments{
\item{exampleData}{A character, should be one of "exp_data", "meth_data", "cnv_data", "amp_gene", "del_gene" ,"label1","label2","zz","exp.p","meth.p","cnv.p"and "pathdata".}
}
\description{
Get the example data of test package for litte trials.
}
\details{
The function getExampleData(ExampleData = "exp.p)") obtains a vector of lncRNAs confirmed to be related with breast cancer. The function getExampleData(ExampleData = "Profile") obtains the expression pr
}
\references{
Subramanian, A., Tamayo, P., Mootha, V.K., Mukherjee, S., Ebert, B.L., Gillette, M.A., Paulovich, A., Pomeroy, S.L., Golub, T.R., Lander, E.S. et al. (2005) Gene set enrichment analysis: a knowledgebased approach for interpreting genome-wide expression profiles. Proc Natl Acad Sci U S A, 102, 15545-15550.
}
