% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimalSubpath.R
\name{opt_subpath}
\alias{opt_subpath}
\title{opt_subpath}
\usage{
opt_subpath(subpathdata, zz, overlap = 0.6)
}
\arguments{
\item{subpathdata}{interested subpathways}

\item{zz}{a vector of z-scores}

\item{overlap}{Overlap ratio of each two pathway genes}
}
\value{
Optimized subpathway:the number of genes shared by any two pathways accounted for less than the Overlap ratio of each pathway genes.
}
\description{
`opt_subpath` Optimize interested subpathways.If the number of genes shared by the two pathways accounted for more than the Overlap ratio of each pathway genes,then combine two pathways.
}
\examples{
zz<-GetExampleData("zzz")
subpathdata<-GetExampleData("subpathdata")
optsubpath<-opt_subpath(subpathdata,zz,overlap=0.6)
}
