% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinep.r
\name{coverp2zscore}
\alias{coverp2zscore}
\title{coverp2zscore}
\usage{
coverp2zscore(pdata)
}
\arguments{
\item{pdata}{A numeric vector of p-values or corrected p-values}
}
\value{
A numeric vector of z_scores
}
\description{
`coverp2zscore` calculate z-scores for p-values
}
\examples{
exp.p<-GetExampleData("exp.p")
meth.p<-GetExampleData("meth.p")
cnv.p<-GetExampleData("cnv.p")
\dontrun{coverp2zscore(exp.p)}
\dontrun{coverp2zscore(meth.p)}
\dontrun{coverp2zscore(cnv.p)}
}
